Update Dockerfile.dapper
This commit updates the versions of packages referred to by the Dockerfile.dapper.

It removes a hack used in the Dockerfile.dapper to fix windows docker image builds and adds a different hack to support the use of cygwin within a Docker container.

It also changes the mirror used for getting the cygwin binary since the old mirror does not appear to exist anymore.